Add files via upload
POP Housing (Photopolymer Open-Source Pressure Housing)
Designed by Breanna Motsenbocker at the University of Rhode Island
Undersea Robotics and Imaging Lab

POP is a compact and customizable deep-sea housing. It has been successfully tested to 4000psi (~2600 msw). The endcap can be modified to allow various passthroughs (bulkhead, ribbon cable, fiber optic, etc.) and/or to attach sensors/other instrumentation to the housing. This way the user can print one main body and switch out endcaps for different applications. 

There are 3 components to POP: the main body, endcap, and collar. The main body has a mount on it that can be used with four 1/4" bolts. The endcap has an O-ring groove made for a dash 226 O-Ring. The collar screws the endcap to the main body and ensures the O-ring is compressed to create a seal. In this repository are the Fusion Files (.f3d), .stl files, and two preform files to show a sample of how to orient each part to keep the O-ring groove/surfaces clean. 

Note: the threads on the collar will have supports, I recommend sanding them down as much as possible and then screwing the collar back and forth on the threads until it wears down.